+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fixed permission errors when other players upload HDC files. [#1494](https://github.com/dmdorman/hero6e-foundryvtt/issues/1494)
 - Improved description for FOCI. [#1521](https://github.com/dmdorman/hero6e-foundryvtt/issues/1521)
 - Main conditional defenses vary (PD/ED), but KB conditional defenses are always PD. [#1518](https://github.com/dmdorman/hero6e-foundryvtt/issues/1518)
+- Uploading HDC prompts to retain BODY, END, and CHARGES. HAP and HeroicIdentity are always retained after upload. [#1505](https://github.com/dmdorman/hero6e-foundryvtt/issues/1518)
 
 ## Version 4.0.8